requirements.txt
# Core ML Libraries
catboost==1.2.8
lightgbm==4.0.0
scikit-learn==1.3.0
xgboost==3.0.2

# Data Processing
pandas==2.3.1
numpy>=2.0.0
dask==2025.5.1

# Feature Store & ML Ops
feast==0.50.0
mlflow==3.1.1
prefect==3.4.8

# Web Framework & API
fastapi==0.116.1
uvicorn[standard]==0.34.0
pydantic>=2.5.0
prometheus-client>=0.19.0
psutil>=5.9.0
python-multipart>=0.0.6

# Data Pipeline
websockets==15.0.1
asyncio-mqtt==0.13.0
redis==4.6.0
psycopg2-binary==2.9.7

# Core Dependencies
pydantic==2.10.6
sqlalchemy==2.0.21
joblib==1.3.2
readchar==4.2.1

# Utilities (if needed by your code)
PyYAML==6.0.2
requests==2.32.4
python-dotenv==1.0.0
click==8.1.7